[Providers] Changes providers for listings & details VCs.
