+++ Kiosk/App/Networking/ArtsyAPI.swift
@@ -319,95 +319,6 @@ extension ArtsyAPI : MoyaTarget {
     }
 }
 
-// MARK: - Provider setup
-
-// (Endpoint<Target>, NSURLRequest -> Void) -> Void
-func endpointResolver() -> MoyaProvider<ArtsyAPI>.RequestClosure {
-    return { (endpoint, closure) in
-        let request: NSMutableURLRequest = endpoint.urlRequest.mutableCopy() as! NSMutableURLRequest
-        request.HTTPShouldHandleCookies = false
-        closure(request)
-    }
-}
-
-class ArtsyProvider<Target where Target: MoyaTarget>: RxMoyaProvider<Target> {
-
-    let online: Observable<Bool>
-
-    init(endpointClosure: MoyaProvider<Target>.EndpointClosure = MoyaProvider.DefaultEndpointMapping,
-        requestClosure: MoyaProvider<Target>.RequestClosure = MoyaProvider.DefaultRequestMapping,
-        stubClosure: MoyaProvider<Target>.StubClosure = MoyaProvider.NeverStub,
-        manager: Manager = Alamofire.Manager.sharedInstance,
-        plugins: [Plugin<Target>] = [],
-        online: Observable<Bool> = connectedToInternetOrStubbing()) {
-
-            self.online = online
-            super.init(endpointClosure: endpointClosure, requestClosure: requestClosure, stubClosure: stubClosure, manager: manager, plugins: plugins)
-    }
-}
-
-struct Provider {
-    private static var endpointsClosure = { (target: ArtsyAPI) -> Endpoint<ArtsyAPI> in
-        
-        var endpoint: Endpoint<ArtsyAPI> = Endpoint<ArtsyAPI>(URL: url(target), sampleResponseClosure: {.NetworkResponse(200, target.sampleData)}, method: target.method, parameters: target.parameters)
-        // Sign all non-XApp token requests
-
-        switch target {
-        case .XApp:
-            return endpoint
-        case .XAuth:
-            return endpoint
-
-        default:
-            return endpoint.endpointByAddingHTTPHeaderFields(["X-Xapp-Token": XAppToken().token ?? ""])
-        }
-    }
-
-    static func APIKeysBasedStubBehaviour(_: ArtsyAPI) -> Moya.StubBehavior {
-        return APIKeys.sharedKeys.stubResponses ? .Immediate : .Never
-    }
-
-    static func DefaultProvider() -> ArtsyProvider<ArtsyAPI> {
-        return ArtsyProvider(endpointClosure: endpointsClosure,
-            requestClosure: endpointResolver(),
-            stubClosure: APIKeysBasedStubBehaviour,
-            plugins: Provider.plugins)
-    }
-    
-    static func StubbingProvider() -> ArtsyProvider<ArtsyAPI> {
-        return ArtsyProvider(endpointClosure: endpointsClosure, requestClosure: endpointResolver(), stubClosure: MoyaProvider.ImmediatelyStub, online: just(true))
-    }
-
-    private struct SharedProvider {
-        static var instance = Provider.DefaultProvider()
-    }
-    
-    static var sharedProvider: ArtsyProvider<ArtsyAPI> {
-        get {
-            return SharedProvider.instance
-        }
-        
-        set (newSharedProvider) {
-            SharedProvider.instance = newSharedProvider
-        }
-    }
-
-    static var plugins: [Plugin<ArtsyAPI>] {
-        return [NetworkLogger<ArtsyAPI>(whitelist: { (target: ArtsyAPI) -> Bool in
-            switch target {
-            case .MyBidPosition: return true
-            default: return false
-            }
-        }, blacklist: { target -> Bool in
-            switch target {
-            case .Ping: return true
-            default: return false
-            }
-        })]
-    }
-}
-
-
 // MARK: - Provider support
 
 func stubbedResponse(filename: String) -> NSData! {
